+++ lib/rubocop/cop/primer/primer_octicon.rb
@@ -2,6 +2,7 @@
 
 require "rubocop"
 require "primer/classify/utilities"
+require "primer/classify/validation"
 
 # :nocov:
 module RuboCop
@@ -53,9 +54,10 @@ module RuboCop
           # check if classes are convertible
           if classes.present?
             system_arguments = ::Primer::Classify::Utilities.classes_to_hash(classes)
+            invalid_classes = (system_arguments[:classes]&.split(" ") || []).select { |class_name| ::Primer::Classify::Validation.invalid?(class_name) }
 
-            # Uses custom classes
-            return if system_arguments[:classes].present?
+            # Uses system argument that can't be converted
+            return if invalid_classes.present?
           end
 
           add_offense(node, message: INVALID_MESSAGE)
@@ -133,7 +135,7 @@ module RuboCop
           # No arguments if they map to the default size
           return if size_attributes.blank? || size_attributes.values.all?(&:blank?)
           # Return mapped argument to `size`
-          return "size: :medium" if size_attributes.values.any? { |val| val == ":medium" }
+          return "size: :medium" if size_attributes.values.any?(":medium")
 
           size_attributes.map do |key, value|
             "#{key}: #{value}"